データ構造を明確化する
